Move etl.design.redshift to etl.dialect.redshift
